Refactor storage backends
